:sparkles: Refactor borax.capp
